Create total_precipation_retrieval.py
This is an example script for retrieving 2 days worth of 3-hourly forecasts of total precipitation for the CARRA-West domain. The 2 days cover a precipitation event on 14-15 September 2017. Notice that the total precipitation field is an accumulated field from the forecast initial time. To avoid spin-up effects - that is too low precipitation in the beginning of the forecasts - 12-hourly precipitation can be obtained by subtracting the 6-hour precipitation forecast from the 18-hour precipitation forecast. 3-hourly precipitation can be obtained similarly from the 6-9, 9-12, 12-15 and 15-18 hour forecasts.