Create surface_flux_retrieve.py
This is an example script for retrieving 1 day's worth of 3-hourly forecasts of surface fluxes for the CARRA-West domain. This day covers a strong Greenland Ice Sheet melting event on 12 July 2012. Notice that the fluxes are accumulated fields from the forecast initial time. 12-hourly fluxes can be obtained by subtracting the 6-hour forecast data from the 18-hour forecast data. 3-hourly fluxes can be obtained similarly from the 6-9, 9-12, 12-15 and 15-18 hour forecasts.